CIS 150 Course Description
This course provides an overview of contemporary information systems, and how they are used in organizations. It also presents an introduction to the discipline of CIS and how information systems can be used to create a competitive advantage in a business environment. Students will develop basic websites including the use of HTML, CSS, and JavaScript.