CIS 199 Course Description
This course introduces object-oriented concepts such as the use of classes, methods, encapsulation, and inheritance, and graphical user interface. The course concentrates on using object-oriented programming to solve simple problems involving input and output. Computer lab sessions are used to reinforce programming concepts. Visual Studio was the tool used to complete each project. C# was the language used in this course.