CIS 410- Course Description
This course explores strategic development of information technology; value chain analysis and its application to information resource management; information systems planning; organizing, staffing, and controlling the deployment of information technology; the development of an IT platform and architecture consistent with organizational structure. This class offers valuable skills for the Management side of business.