CIS 420 - Couse Description
This course focuses on the detailed design and implementation phases of the system development life cycle, including user acceptance testing, test planning, design reviews, and change procedures. Specifications are used to implement, test, and install a working version of an information system. System deployment emphasizes a web-based architecture. A prototyping approach is taken to develop and test the system in an iterative manner. Students are grouped into project teams, and each team member accepts the task assignments necessary to deliver the information system prototype.